Rename users table to lifters
`Users` has a bit of a negative connotation for me, so in my app,
they're called lifters to be tightly coupled to the domain.

"Tightly coupled" to the domain sounds like something I shouldn't be
doing at the database level, but again, it's my app.

We also make the following changes from the original SQL script:

- Rename `name` to `user_name` (I plan of auto-generating user names)
- Remove unique constraint from `user_name` column (just the email needs
  to be unique)